Remove closeCloser Function
The function was a left over of the initial code. It closed io.ReadCloser
but did throw a "panic" if closer.Close() caused an error. SDK should not
panic and therefore, the function needed to be refactored.

The current solution could be better, but I opted for a "simple" solution
without using defer with an anonymous function and named return values
like the AWS SDK. This would probably be the "better" solution but is
also harder to read for newcomers.